+++ examples/test_scheduler.ex
@@ -1,0 +1,9 @@
+defmodule TestScheduler do
+  use Bonny.Server.Scheduler, name: "test-scheduler"
+
+  def select_node_for_pod(_pod, nodes) do
+    nodes
+    |> Enum.take(1)
+    |> List.first()
+  end
+end