+++ test/bonny/server/reconciler_test.exs
@@ -1,72 +1,47 @@
 defmodule Bonny.Server.ReconcilerTest do
   @moduledoc false
   use ExUnit.Case, async: true
-  alias Bonny.Server.Reconciler
-  import ExUnit.CaptureLog
 
-  defmodule TestReconciler do
-    use Bonny.Server.Reconciler, frequency: 15
+  alias Bonny.Server.Reconciler, as: MUT
 
-    @impl true
-    def reconcile(%{} = resource) do
-      Agent.update(TestReconcilerCache, fn resources -> [resource | resources] end)
-      :ok
-    end
+  defmodule K8sMock do
+    import K8s.Test.HTTPHelper
 
-    @impl true
-    def reconcile_operation() do
-      K8s.Client.list("example.com/v1", :foos)
-    end
-  end
+    def conn(), do: Bonny.K8sMock.conn(__MODULE__)
 
-  defmodule TestReconcilerErrors do
-    use Bonny.Server.Reconciler, frequency: 15
+    def request(:get, "apis/example.com/v1/foos", _, _, opts) do
+      limit = get_in(opts, [:params, :limit])
 
-    @impl Bonny.Server.Reconciler
-    def reconcile(%{} = resource) do
-      Agent.update(TestReconcilerCacheErr, fn resources -> [resource | resources] end)
-      :ok
-    end
-
-    @impl Bonny.Server.Reconciler
-    def reconcile_operation() do
-      K8s.Client.list("example.com/v1", :errors)
-    end
-  end
-
-  test "schedule/2 sends `:run` after delay" do
-    Reconciler.schedule(self(), 1)
-    assert_receive :run, 2_000
-  end
-
-  describe "run/1" do
-    test "happy path" do
-      Agent.start_link(fn -> [] end, name: TestReconcilerCache)
-      Reconciler.run(TestReconciler)
-      Process.sleep(10)
-
-      resources = Agent.get(TestReconcilerCache, fn resources -> resources end)
-
-      names =
-        resources
-        |> Enum.map(fn %{"name" => name} -> name end)
-        |> Enum.sort()
-
-      assert names == ["bar", "foo"]
-    end
-
-    test "Handles a stream with errors" do
-      fun = fn ->
-        Agent.start_link(fn -> [] end, name: TestReconcilerCacheErr)
-        Reconciler.run(TestReconcilerErrors)
-        Process.sleep(10)
-
-        resources = Agent.get(TestReconcilerCacheErr, fn resources -> resources end)
-
-        assert [%{"name" => "bar"}] == resources
+      case limit do
+        10 -> render(%{"items" => [%{"name" => "foo"}, %{"name" => "bar"}]})
+        1 -> render(%{"metadata" => %{"resourceVersion" => "1337"}})
       end
-
-      assert capture_log(fun) =~ "bonny.reconciler.fetch.failed"
     end
+  end
+
+  defmodule TestController do
+    @behaviour MUT
+
+    def reconcile(resource) do
+      # Process runing test was registered under the name of its module
+      send(Bonny.Server.ReconcilerTest, {:name, resource["name"]})
+
+      :ok
+    end
+  end
+
+  setup do
+    [conn: __MODULE__.K8sMock.conn()]
+  end
+
+  @tag wip: true
+  test "reconciler returns a prepared stream that calls the reconcile function when run", %{conn: conn} do
+    Process.register(self(), __MODULE__)
+
+    operation = K8s.Client.list("example.com/v1", :foos)
+    MUT.get_stream(TestController, conn, operation) |> Stream.run()
+
+    assert_receive {:name, "foo"}
+    assert_receive {:name, "bar"}
   end
 end