+++ test/bonny/server/watcher_test.exs
@@ -1,96 +1,111 @@
 defmodule Bonny.Server.WatcherTest do
   @moduledoc false
   use ExUnit.Case, async: true
-  alias Bonny.Server.Watcher
-  doctest Bonny.Server.Watcher
 
-  defmodule TestWatcher do
-    use Bonny.Server.Watcher
+  alias Bonny.Server.Watcher, as: MUT
 
-    @impl true
-    def watch_operation() do
-      K8s.Client.list("example.com/v1", :watchers)
+  defmodule K8sMock do
+    require Logger
+
+    import K8s.Test.HTTPHelper
+
+    def conn(), do: Bonny.K8sMock.conn(__MODULE__)
+
+    def request(:get, "apis/example.com/v1/widgets", _body, _headers, opts) do
+      case get_in(opts, [:params, :watch]) do
+        true ->
+          pid = Keyword.fetch!(opts, :stream_to)
+
+          send_object(pid, %{
+            "type" => "ADDED",
+            "object" => %{
+              "apiVersion" => "v1",
+              "kind" => "Namespace",
+              "metadata" => %{"name" => "foo", "resourceVersion" => "11"}
+            }
+          })
+
+          send_object(pid, %{
+            "type" => "ADDED",
+            "object" => %{
+              "apiVersion" => "v1",
+              "kind" => "Namespace",
+              "metadata" => %{"name" => "bar", "resourceVersion" => "12"}
+            }
+          })
+
+          send_object(pid, %{
+            "type" => "MODIFIED",
+            "object" => %{
+              "apiVersion" => "v1",
+              "kind" => "Namespace",
+              "metadata" => %{"name" => "foo", "resourceVersion" => "13"}
+            }
+          })
+
+          send_object(pid, %{
+            "type" => "DELETED",
+            "object" => %{
+              "apiVersion" => "v1",
+              "kind" => "Namespace",
+              "metadata" => %{"name" => "foo", "resourceVersion" => "14"}
+            }
+          })
+
+
+          {:ok, %HTTPoison.AsyncResponse{id: make_ref()}}
+
+        nil ->
+          render(%{"metadata" => %{"resourceVersion" => "10"}})
+      end
     end
 
-    @impl true
+    def request(_method, _url, _body, _headers, _opts) do
+      Logger.error("Call to #{__MODULE__}.request/5 not handled: #{inspect(binding())}")
+      {:error, %HTTPoison.Error{reason: "request not mocked"}}
+    end
+  end
+
+
+  defmodule TestController do
+    @behaviour MUT
+
     def add(resource) do
-      track_event(:add, resource)
+      # Process runing test was registered under the name of its module
+      send(Bonny.Server.WatcherTest, {:add, resource["metadata"]["name"]})
+      :ok
     end
 
-    @impl true
     def modify(resource) do
-      track_event(:modify, resource)
+      # Process runing test was registered under the name of its module
+      send(Bonny.Server.WatcherTest, {:modify, resource["metadata"]["name"]})
+      :ok
     end
 
-    @impl true
     def delete(resource) do
-      track_event(:delete, resource)
-    end
-
-    @spec track_event(atom, map) :: :ok
-    def track_event(type, resource) do
-      event = {type, resource}
-      Agent.update(TestWatcherCache, fn events -> [event | events] end)
+      # Process runing test was registered under the name of its module
+      send(Bonny.Server.WatcherTest, {:delete, resource["metadata"]["name"]})
+      :ok
     end
   end
 
-  @tag :wip
-  test "watch/3" do
-    Agent.start_link(fn -> [] end, name: TestWatcherCache)
-    {:ok, pid} = TestWatcher.start_link()
-    rv = "3"
-
-    Watcher.watch(TestWatcher, rv, pid)
-    Process.sleep(500)
-
-    events = Agent.get(TestWatcherCache, fn events -> events end)
-    refute events == []
+  setup do
+    [conn: __MODULE__.K8sMock.conn()]
   end
 
-  describe "dispatch/2" do
-    test "dispatches ADDED events to the given module's handler function" do
-      evt = event("ADDED")
-      Watcher.dispatch(evt, Whizbang)
+  test "watcher returns a prepared stream that calls the add/modify/delete functions for each event", %{conn: conn} do
+    Process.register(self(), __MODULE__)
 
-      # Professional.
-      :timer.sleep(100)
-      assert [event] = Whizbang.get(:added)
-      assert %{"apiVersion" => "example.com/v1"} = event
-    end
+    operation = K8s.Client.list("example.com/v1", :widgets)
+    stream = MUT.get_stream(__MODULE__.TestController, conn, operation) |> Stream.take(4)
+    Task.async(fn ->
+      Stream.run(stream)
+    end)
+    |> Task.await()
 
-    test "dispatches MODIFIED events to the given module's handler function" do
-      evt = event("MODIFIED")
-      Watcher.dispatch(evt, Whizbang)
-
-      # Professional.
-      :timer.sleep(100)
-      assert [event] = Whizbang.get(:modified)
-      assert %{"apiVersion" => "example.com/v1"} = event
-    end
-
-    test "dispatches DELETED events to the given module's handler function" do
-      evt = event("DELETED")
-      Watcher.dispatch(evt, Whizbang)
-
-      # Professional.
-      :timer.sleep(100)
-      assert [event] = Whizbang.get(:deleted)
-      assert %{"apiVersion" => "example.com/v1"} = event
-    end
-  end
-
-  defp event(type) do
-    %{
-      "object" => %{
-        "apiVersion" => "example.com/v1",
-        "kind" => "Widget",
-        "metadata" => %{
-          "name" => "test-widget",
-          "namespace" => "default",
-          "resourceVersion" => "705460"
-        }
-      },
-      "type" => type
-    }
+    assert_received {:add, "foo"}
+    assert_received {:add, "bar"}
+    assert_received {:modify, "foo"}
+    assert_received {:delete, "foo"}
   end
 end